refactoring : api URL 변경, axios 부분 변경
