+++ src/hooks/use-current-user.ts
@@ -4,7 +4,7 @@ export function useCurrentUser() {
   const [user, setUser] = useState<{ userId: number } | null>(null)
 
   useEffect(() => {
-    fetch("http://localhost:8080/api/user/me", {
+    fetch(process.env.REACT_APP_API_BASE_URL + "/user/me", {
       credentials: "include", // 쿠키 기반 로그인이라면 반드시 필요
     })
       .then((res) => res.json())
